anae_venture_admin/backend/requirements.txt
django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
pgvector==0.3.3
pillow==10.4.0
django-extensions==3.2.3
ollama==0.3.3
redis==5.0.8
django-ratelimit==4.1.0
drf-spectacular==0.27.2
Faker==30.0.0
django-coreapi==2.3.0
channels
celery 
daphne
channels_redis
django-celery-beat
django-celery-results
python-decouple


